radio-llama/llamapi/utils_romain/requirements.txt
spotdl
langchain
langchain_core
langchain_mistralai
ffmpeg-python